ssa: Move OrdLabel and utils out, Fix all trees FINALLY
